requirements.txt
streamlit
google-generativeai
thefuzz
python-dotenv
pandas
pillow
